requirements.txt
gum-ai
python-dotenv
ics
aiohttp